requirements.txt
pandas==1.2.4
bokeh==2.3.2
numpy==1.20.1
streamlit==0.89.0
pydeck==0.7.0
Pillow==8.3.2
